backend/requirements.txt
fastapi==0.115.5
uvicorn==0.32.0
pydantic==2.9.2
chromadb==0.5.15
pypdf==4.3.1
pillow==10.4.0
pytesseract==0.3.13
pymupdf==1.24.10
